code/pyproject.toml
[project]
name = "fsl-rainfall"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2",
    "grpcio>=1.62",
    "grpcio-tools>=1.62",
    "numpy>=1.26",
    "pandas>=2.2",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "requests>=2.31",
    "scikit-learn>=1.4",
    "tqdm>=4.66",
    "uo-pyfetch",
    "pyarrow>=23.0.1",
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "data", "models", "client", "server", "sim", "utils", "experiments"]

[tool.uv.sources]
uo-pyfetch = { url = "https://api.v2.urbanobservatory.ac.uk/lib/uo-pyfetch.tar.gz" }
